--- src/About.jsx
@@ -1,22 +1,0 @@
-import React from "react";
-
-function About(props) {
-  return (
-    <div className="about-title">
-      <h1 className="title">I am a software developer</h1>
-      <p>
-        Mi nombre es Irving Rabanales y he trabajado en varios projectos. Puedo
-        desarrollar cualquier applicacion hibrida o nativa para tu negocio.{" "}
-        <br /> Durante este tiempo en el cual la tecnologia bien con un boom
-        total ayudarte a darle esa personificacion tecnologica.
-      </p>
-      <h2>"I" solutions</h2>
-      <p>
-        Aqui es donde mi compañia te podra dar todo lo necesario para que tu
-        negocio triunfe en el mercado tecnologico con asesoria personalizada.
-      </p>
-    </div>
-  );
-}
-
-export default About;
