+++ src/App.css
@@ -1,5 +1,5 @@
 .navigation {
-  background-color: #fff;
+  background-color: rgb(228, 228, 228);
 }
 
 div.navigation-sub {
