--- src/Articles.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Articles(props) {
-  return <div></div>;
-}
-
-export default Articles;
