--- src/Projects.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import Thumbnail from "./Thumbnail";
-
-function Projects(props) {
-  return (
-    <div className="projects-area">
-      <h1 className="projects-title">"I" Solutions!</h1>
-      
-    </div>
-  );
-}
-
-export default Projects;
