--- src/Thumbnail.jsx
@@ -1,19 +1,0 @@
-import React from "react"; // Import the Component component from React
-import { Link } from "react-router-dom"; // Import the Link component
-import "./App.css";
-
-function Thumbnail(props) {
-  return (
-    <div className="project">
-      <Link to={props.link}>
-        <div className="project-image">
-          <img src={props.images} alt="Project Image" />
-        </div>
-        <div className="project-title">{props.title}</div>
-        <div className="project-category">{props.title}</div>
-      </Link>
-    </div>
-  );
-}
-
-export default Thumbnail;
