+++ src/components/Articles.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Articles(props) {
+  return <div></div>;
+}
+
+export default Articles;
