+++ src/components/Projects.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import Avatar from "@material-ui/core/Avatar";
+
+export default function Projects() {
+  return (
+    <div className="projects-area">
+      <h1 className="projects-title">"I" Solutions!</h1>
+    </div>
+  );
+}
